Refactor tree component
